% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Returns meta data on MSMT files}
\usage{
get_locations(location = NULL)
}
\arguments{
\item{location}{specify path to the folder where MSMT data can be found. The function uses recursive search, so a higher level directory can be specified as well. If left blank, the function uses the current working directory.}
}
\value{
a tibble with all sheets, specifying their location, names, lowercased name, form they are from, year and frequency of occurences in the folders.
}
\description{
`get_locations` returns a meta data map of relevant MSMT files in a given folder.
}
\examples{
get_locations("E:/EDU/Data/data/MSMT/")
}
