% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_names.R
\name{get_table_names}
\alias{get_table_names}
\title{Creates variable name matrix for easier use of form indices}
\usage{
get_table_names(
  table_ind,
  row_inds,
  col_inds,
  extra_col = NULL,
  extra_row = NULL,
  omits = NULL,
  prefix = "r",
  column_names = NULL,
  rows_names = NULL
)
}
\arguments{
\item{table_ind}{A single character indicating the table prefix}

\item{row_inds}{A numeric vector indicating requested row indices. If some rows are indexed using a letter, use `extra_row` parameter}

\item{col_inds}{A numeric vector indicating requested column indices. If some column are indexed using a letter, use `extra_col` parameter}

\item{extra_col}{A character vector indicating columns indexed with a character. If the index starts with 0, include it in the character vector, e.g. c("07a","07b")}

\item{extra_row}{A character vector indicating rows indexed with a character. If the index starts with 0, include it in the character vector, e.g. c("07a","07b")}

\item{omits}{If you would like to replace some cells with NA, include a list of matrix indices (e.g. list(c(1,1),c(1,2))) of requested cells. Cell indices reflect values of all entered row/column index names, including extra_ as they would appear in the form}

\item{prefix}{As variable names start with "r", prefix defaults to that value, but can be replaced}

\item{column_names}{A character vector (optional) where names of columns can be manually assigned. Order of columns will reflect column indices sorted as they would appear in the form, including those from extra_col}

\item{rows_names}{A character vector (optional) where names of rows can be manually assigned. Order of rows will reflect sorted row indices sorted as they would appear in the form, including those from extra_row}
}
\value{
a matrix of cell names that can be used to call indexed groups of cells from MSMT data
}
\description{
`get_table_names` facilitates the reconstruction of cell names from MSMT forms. The format of cell indices takes the following format: xxyyz. "xx" is the table prefix common to all values in table. yy indicates the row and y (usually only one, but can be longer) stands for the column.
}
\examples{
get_table_names(table_ind = "3B", row_inds = 1:14, col_inds = 2:10, extra_col = c("10a","10b"))
}
