% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_availability.R
\name{get_variable_availability}
\alias{get_variable_availability}
\title{Show the availability of requested variables across selected forms over years}
\usage{
get_variable_availability(
  variables,
  map = NULL,
  forms = NULL,
  print_plots = FALSE
)
}
\arguments{
\item{variables}{A character vector with requires variable names in lowercase.}

\item{map}{A folder map retrieved from map_folders. If left empty, the function will search the working directory for a map file. Can be also used to specify file path to the map file.}

\item{forms}{A character vector with forms of interest, in the following form c("v11","v12") etc., where the number specifies form number. If blank, all forms are used. For form numbers, consult forms_codes().}

\item{print_plots}{Logical indicating whether the function should print plots. Plots will be available in output even if not printed.}
}
\value{
a list of tibbles with requested variables and identifiers, and a tibble with information about used sheets. To join into a single tibble, use bind_rows() on the tibble list.
}
\description{
Using a map of variables, `get_variable_availability` returns a list of tibbles showing the availability of requested variables in time
}
\examples{
availability <- get_variable_availability(variables = c("r01010","r01011"), map = my_map, forms = "v03")
availability$plots$v03
}
