% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Return a list of tibbles with requested MSMT data}
\usage{
get_variables(
  variables,
  joint = c("red_izo", "izo", "p_izo", "izonew"),
  map = NULL,
  forms = NULL,
  years = NULL
)
}
\arguments{
\item{variables}{A character vector with requires variable names in lowercase.}

\item{joint}{A character vector including case identifiers to look for in the data. Uses a vector of default identifiers.}

\item{map}{A folder map retrieved from map_folders. If left empty, the function will search the working directory for a map file. Can be also used to specify file path to the map file.}

\item{forms}{A character vector with forms of interest, in the following form c("v11","v12") etc., where the number specifies form number. If blank, all forms are used. For form numbers, consult forms_codes().}

\item{years}{A character vector with years of interest, in the following form c("12", "13") etc., where the number specify the last two numerics of a given year. If blank, all years are used.}
}
\value{
a list of tibbles with requested variables and identifiers, and a tibble with information about used sheets. To join into a single tibble, use bind_rows() on the tibble list.
}
\description{
Using a map of variables, `get_variables` returns tibbles with required variables and identifiers from specified years and forms.
}
\examples{
mymap <- get_locations("E:/EDU/Data/data/MSMT/")
get_variables(variables = c("r01010","r01011"), map = my_map, forms = "v03", years = c("12","13"))
}
