% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varnames.R
\name{get_varnames}
\alias{get_varnames}
\title{Returns names of variables in MSMT files}
\usage{
get_varnames(location = NULL, forms = NULL, years = NULL)
}
\arguments{
\item{location}{A tibble from get_locations function. If left blank, uses get_location with no specified parameters.}

\item{forms}{A character vector with forms of interest, in the following form c("v11",v"12") etc., where the number specifies form number. For form numbers, consult forms_codes().}

\item{binary:}{print boolean (TRUE) or counts (FALSE) for repeated observations}
}
\value{
a list of values indicating files with repeated rows
}
\description{
`get_varnames`Uses the output of get_locations function and returns the list of all variable names in available data.
}
\examples{
mymap <- get_locations("E:/EDU/Data/data/MSMT/")
get_varnames(location = mymap, forms = "v11")
}
