% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_folders.R
\name{map_folders}
\alias{map_folders}
\title{Creates a map of MSMT data for future reference}
\usage{
map_folders(location = NULL, save_map = TRUE, return_map = FALSE)
}
\arguments{
\item{location}{A single character with file path to the directory containing MSMT data. If left blank, working directory is used as a default.}

\item{save_map}{If TRUE, the function saves a map in the working directory. Can be set to FASLE, or replaced with a single character specifying path to the folder to save the map.}

\item{return_map}{If TRUE, the function returns the map, so it can be used in the workspace.}
}
\value{
A list containing locations of excel sheets and variable names occurring in the sheets.
}
\description{
`map_folders` maps sheets containing form data in MSMT folder: their locations and variables. The function saves the map in an .RDATA format for future use, but can be also used to create a workspace object.
}
\examples{
map_folders(location = "E:/EDU/Data/data/MSMT/")
}
